cart/requirements.txt
Flask==1.0.2
pymongo==3.9.0
requests==2.21.0
Flask_Cors==3.0.8
pika==1.1.0
python-dotenv==0.12.0
shortuuid==1.0.1
dnspython
